GUI icons, blocked requests: considering "default policy"
First, the flag. So far the flag turned red as soon as *anything* was
blocked. With this commit, the flag does turn red *only* if at least one
of the blocked requests were blocked by the default policy ("Block
requests by default" setting)

Second, in the menu in the list of blocked requests, there are now icons.
A question mark means that the default policy applied (see above).
In any other case, a "blocked" sign is shown, this is, there's a rule
(user or subscription rule).

In summary, when "Block requests by default" is set and the flag is red,
you will find question marks in the menu.

This commit affects (and maybe partly solves) issues #98, #368, #389,

on the technical side, I renamed the following functions:
[*] CheckRequestResult:  isAllowed()  →  allowRulesExist()
[*] CheckRequestResult:  isDenied()  →  denyRulesExist()
[*] RequestPolicyService:  _isAllowedByDefaultPolicy() → _checkByDefaultPolicy()
    ---> now returns a CheckRequestResult object instead of true/false.

…new functions:
[*] RequestPolicyService.checkRedirect(originUri, destinationUri)
    ---> same as isAllowedRedirect(), but returns a CheckRequestResult
         object instead of true/false (contains more information).
[*] requestpolicy.menu._extractRequestProperties

a new Class:
[*] Destination