[menu] show numRequests also for origin & otherOrigins
this feature is now complete

also, now the flag will be shown in front of the origin(s)
that flag is specific to that origin, so it might be that the "main"
origin has no red flag, but an other origin.

There is now a new file: GUILocation.jsm. It contains several classes
which are used in menu.js. The "Destination" class has been renamed to
"GUIDestination" which is now located in GUILocation.jsm.