bug found: not always display the redir.notif.bar!
Currently the redirection notification bar is displayed if *any*
request in a Tab causes a redirect. But instead, it should only
be displayed in case of an explicit user interaction, i.e an url
is entered in the location bar or a link has been clicked.

A mozmill test has already been created, but it's disabled for
now: tests/mozmill/tests/testRedirect/testInlineRedirect.js