Add `Environment` class, replaces BootstrapManager
This commit introduces the Environment class. There is also the
`ProcessEnvironment` module, which simply creates one Environment instance
per process that is available for all modules.

Environment implements the main functionality of BootstrapManager, but as it's
a class, more instances can be created, such as "window" environments. The
Environment class simplifies greatly the startup and shutdown processes for
both the main process and frame scripts, as well as for content windows running
in the main process such as the preferences web pages.

By the way, ObserverManager profits a lot from this change, as it dosn't
need to detect the process type itself anymore.