introducing EnvironmentManager, fixing many bugs
- The EnvironmentManager has been introduced. Documentation can
be found in the Development Wiki.

- Many bugs in respect of RP's shutdown (and startup) have been
solved.

- bootstrap.js now starts up the EnvironmentManager instead of
ProcessEnvironment. This means that ProcessEnvironment is no
longer responsible for essential tasks on startup/shutdown.

- The Environment class now supports "Levels": Essential,
Backend, Interface and UI. All startup and shutdown functions
are assigned to one of those levels. For both the startup and
the shutdown process the levels are processed in the right
order.

- Two new managers have been added, helping with startup and
shutdown:
- ManagerForMessageListeners
- ManagerForEventListeners
Both of them add/remove listeners on startup/shutdown; Every
instance of those classes is bound to an environment.