eliminate framescript DocManager, which was buggy
`DocManager` was only used in the "DOMContentLoaded" callback.
That callback now notifies the chrome process *synchronously*,
so it's not needed to hold a reference to the document.

In order to let the page load as fast as possible, calling
`indicateBlockedVisibleObjects()` is done async. The indicators
might appear a little later than the page itself.