integrate EnvMan into Environment, et cetera
This commit removes the EnvironmentManager introduced by
commit `75aca76`. The managing-functionality has been merged
into the `Environment` class itself, which means that each
environment manages its inner environments. The EnvMan's
special startup- and shutdown-tasks can now be found in
`ProcessEnvironment` which is an instance of one of the
following classes:
- ParentProcessEnvironment
- ChildProcessEnvironment
Those two classes aren't exported, as each process needs only
one instance of one of the two classes.
Further information will be documented in the wiki.

Another change, which is even more significant, is a bugfix
introduced as well by commit `75aca76`. That commit made the
assumption that if a framescript is in a child process, that
framescript will be the only framescript in that process.
THIS IS WRONG. Usually there are several framescripts in the
same process. The opposite should be seldomly the case. The
solution was to