use the globalMM for loading frame scripts
Before this commit, frame scripts have been loaded via the
window's message manager. There was a strange bug, and
using the global message manager instead fixed the bug.

The issue was that in some tabs some event listeners haven't
been registered correctly. As an example, the event
"DOMContentLoaded" wasn't observed in some tabs so that in
that tabs the red flag wasn't displayed for blocked content.