show a notice when multiple RP versions are active
When RP is installed/enabled, it will check if there are other
versions of RP installed and enabled. RP will also observe addon
install/enable actions, so that it recognizes when another RP
version gets active.

In any of the above cases, a new tab will be opened, which will
show a notice that there shouldn't be multiple instances of RP
active at the same time.

A Marionette test suite ensures the desired functionality.
Thanks to the "addons" library, this can be accomplished quite
easily.