[ref] overlay: prevent AMO's `setTimeout` warning
The message of the AMO Validator was:

```
`setTimeout` called in potentially dangerous manner

Warning: In order to prevent vulnerabilities, the `setTimeout`
and `setInterval` functions should be called only with function
expressions as their first argument.

Severity for automated signing: high
Suggestions for passing automated signing:

Please do not ever call `setTimeout` or `setInterval` with
string arguments. If you are passing a function which is not
being correctly detected as such, please consider passing a
closure or arrow function, which in turn calls the original
function.
```