[tst][add] RP Puppeteer "your_policy" library
This library allows to control the elements on the "Your Policy"
settings page. Some parts of this library have been taken from
the old `testRuleWithSchemeOnly.js` Mozmill test.

The tests for the "Your Policy" library check that the "origin"
and "destination" strings in the table are correct for all
possible combinations of scheme, host and port. In case of a
rule which only specifies a scheme, the test assumes the
string `scheme "http"` if the scheme is "http". Previously
the string was `http:*`, which however could be confused
with `*://http:*`.