[fix] http-on-examine-response: top-level doc detection
On E10s there was an error in the "http-on-examine-response"
observer, specifically in the function
`eventuallyShowRedirectNotification()`. The function checks if
the load corresponding to the request is a top-level document
request. If true, the redirection notification is shown.

The function used the DocShell's "busy" flags, but the DocShell
lives in the content process. The solution is to use the
channel's load flags instead. It's compatible with E10s and it's
even less code.

However, the redirection notification still isn't shown on
http://www.maindomain.test/redirect-http-location-header-png.php
because there are more issues. See also #721.